Created controller for placing.
